fix: moves components folder out from app folder
